requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
google-cloud-bigquery==2.34.2
pytest